+++ src/irregint.h
@@ -16,6 +16,6 @@ gboolean from_uint64_to_uint48(guchar * res, guint64 _48);
 void from_uint48_to_uint64(guchar * input, guint64 * _48);
 
 void from_8x_bool_to_uint8(guint8 * res, gboolean _bools[8]);
-void from_uint8_to_8x_bool(guint input, gboolean _bools[8]);
+void from_uint8_to_8x_bool(guint8 input, gboolean _bools[8]);
 
 #endif
